+++ bsp_tool/branches/ace_team/__init__.py
@@ -1,0 +1,5 @@
+"""Creators of Zeno Clash"""
+from . import zeno_clash
+
+
+scripts = [zeno_clash]