+++ bsp_tool/branches/nexon/__init__.py
@@ -3,6 +3,7 @@ They have worked with both Valve & Respawn's Source Engine (CS:O & TF:O [cancell
 from . import cso2
 from . import cso2_2018
 from . import vindictus
+from . import vindictus69
 
 
-scripts = [cso2, cso2_2018, vindictus]
+scripts = [cso2, cso2_2018, vindictus, vindictus69]