+++ bsp_tool/branches/outerlight/__init__.py
@@ -1,0 +1,5 @@
+"""Creators of The Ship & Bloody Good Time"""
+from . import outerlight
+
+
+scripts = [outerlight]