+++ bsp_tool/branches/outerlight/outerlight.py
@@ -1,10 +1,6 @@
 import enum
-from typing import List
 
-from .. import base
-from .. import vector
 from ..valve import orange_box
-from ..valve import sdk_2013
 from ..valve import source
 
 
@@ -12,8 +8,8 @@ FILE_MAGIC = b"VBSP"
 
 BSP_VERSION = 20
 
-GAME_PATHS = {"Fairy Tale Busters": "Merubasu/shadowland"}
-# NOTE: supported by Jabroni Brawl 3
+GAME_PATHS = {"Bloody Good Time": "Bloody Good Time/pm",
+              "The Ship": "The Ship/ship"}  # same for Single Player & Tutorial
 
 GAME_VERSIONS = {GAME_NAME: BSP_VERSION for GAME_NAME in GAME_PATHS}
 
@@ -97,41 +93,8 @@ LumpHeader = source.LumpHeader
 
 
 # classes for special lumps, in alphabetical order:
-class StaticPropv11(base.Struct):  # sprp GAME LUMP (LUMP 35) [version 11]
-    """https://github.com/ValveSoftware/source-sdk-2013/blob/master/sp/src/public/gamebspfile.h#L186"""
-    # older spec than sdk_2013's? might need to rethink the sdk_2013 script
-    origin: List[float]  # origin.xyz
-    angles: List[float]  # origin.yzx  QAngle; Z0 = East
-    model_name: int  # index into GAME_LUMP.sprp.model_names
-    first_leaf: int  # index into GAME_LUMP.sprp.leaves -> Leaf lump
-    num_leafs: int  # number of Leaves after first_leaf this StaticProp is in
-    # incorrect:
-    solid_mode: int  # collision flags enum
-    flags: int  # other flags
-    skin: int  # index of this StaticProp's skin in the .mdl
-    fade_distance: List[float]  # min & max distances to fade out
-    lighting_origin: List[float]  # xyz position to sample lighting from
-    forced_fade_scale: float  # relative to pixels used to render on-screen?
-    cpu_level: List[int]  # min, max (-1 = any)
-    gpu_level: List[int]  # min, max (-1 = any)
-    diffuse_modulation: List[int]  # RGBA 32-bit colour
-    flags_2: int  # values unknown
-    # correct:
-    scale: float
-    __slots__ = ["origin", "angles", "name_index", "first_leaf", "num_leafs",
-                 "solid_mode", "flags", "skin", "fade_distance", "lighting_origin",
-                 "forced_fade_scale", "cpu_level", "gpu_level", "diffuse_modulation",
-                 "flags_2", "scale"]
-    _format = "6f3H2Bi6f8BIf"
-    _arrays = {"origin": [*"xyz"], "angles": [*"yzx"], "fade_distance": ["min", "max"],
-               "lighting_origin": [*"xyz"], "cpu_level": ["min", "max"],
-               "gpu_level": ["min", "max"], "diffuse_modulation": [*"rgba"]}
-    _classes = {"origin": vector.vec3, "solid_mode": source.StaticPropCollision, "flags": source.StaticPropFlags,
-                "lighting_origin": vector.vec3}  # TODO: angles QAngle, diffuse_modulation RBGExponent
-
-
-class GameLump_SPRPv11(sdk_2013.GameLump_SPRPv11):  # sprp GAME LUMP (LUMP 35) [version 11]
-    StaticPropClass = StaticPropv11
+# TODO: GameLump_SPRPv5
+# TODO: GameLump_SPRPv6
 
 
 # {"LUMP_NAME": {version: LumpClass}}
@@ -144,7 +107,7 @@ SPECIAL_LUMP_CLASSES = orange_box.SPECIAL_LUMP_CLASSES.copy()
 GAME_LUMP_HEADER = orange_box.GAME_LUMP_HEADER
 
 # {"lump": {version: SpecialLumpClass}}
-GAME_LUMP_CLASSES = {"sprp": {7: GameLump_SPRPv11}}
+GAME_LUMP_CLASSES = dict()
 
 
 methods = [*orange_box.methods]