+++ bsp_tool/branches/valve/source_filmmaker.py
@@ -1,0 +1,112 @@
+import enum
+
+from ..valve import sdk_2013
+from ..valve import source
+
+
+FILE_MAGIC = b"VBSP"
+
+BSP_VERSION = 21
+
+GAME_PATHS = {"Source Filmmaker": "SourceFilmmaker/game/tf"}
+
+GAME_VERSIONS = {GAME_NAME: BSP_VERSION for GAME_NAME in GAME_PATHS}
+
+
+class LUMP(enum.Enum):
+    ENTITIES = 0
+    PLANES = 1
+    TEXTURE_DATA = 2
+    VERTICES = 3
+    VISIBILITY = 4
+    NODES = 5
+    TEXTURE_INFO = 6
+    FACES = 7  # version 1
+    LIGHTING = 8  # version 1
+    OCCLUSION = 9  # version 2
+    LEAVES = 10  # version 1
+    FACE_IDS = 11  # TF2 branch, for mapping debug & detail prop seed
+    EDGES = 12
+    SURFEDGES = 13
+    MODELS = 14
+    WORLD_LIGHTS = 15
+    LEAF_FACES = 16
+    LEAF_BRUSHES = 17
+    BRUSHES = 18
+    BRUSH_SIDES = 19
+    AREAS = 20
+    AREA_PORTALS = 21
+    UNUSED_22 = 22
+    UNUSED_23 = 23
+    UNUSED_24 = 24
+    UNUSED_25 = 25
+    DISPLACEMENT_INFO = 26
+    ORIGINAL_FACES = 27
+    PHYSICS_DISPLACEMENT = 28
+    PHYSICS_COLLIDE = 29
+    VERTEX_NORMALS = 30
+    VERTEX_NORMAL_INDICES = 31
+    DISPLACEMENT_LIGHTMAP_ALPHAS = 32  # deprecated / X360 ?
+    DISPLACEMENT_VERTICES = 33
+    DISPLACEMENT_LIGHTMAP_SAMPLE_POSITIONS = 34
+    GAME_LUMP = 35
+    LEAF_WATER_DATA = 36
+    PRIMITIVES = 37
+    PRIMITIVE_VERTICES = 38  # deprecated / X360 ?
+    PRIMITIVE_INDICES = 39
+    PAKFILE = 40
+    CLIP_PORTAL_VERTICES = 41
+    CUBEMAPS = 42
+    TEXTURE_DATA_STRING_DATA = 43
+    TEXTURE_DATA_STRING_TABLE = 44
+    OVERLAYS = 45
+    LEAF_MIN_DIST_TO_WATER = 46
+    FACE_MACRO_TEXTURE_INFO = 47
+    DISPLACEMENT_TRIS = 48
+    PHYSICS_COLLIDE_SURFACE = 49  # deprecated / X360 ?
+    WATER_OVERLAYS = 50  # deprecated / X360 ?
+    LEAF_AMBIENT_INDEX_HDR = 51
+    LEAF_AMBIENT_INDEX = 52
+    LIGHTING_HDR = 53  # version 1
+    WORLD_LIGHTS_HDR = 54
+    LEAF_AMBIENT_LIGHTING_HDR = 55  # version 1
+    LEAF_AMBIENT_LIGHTING = 56  # version 1
+    XZIP_PAKFILE = 57  # deprecated / X360 ?
+    FACES_HDR = 58  # version 1
+    MAP_FLAGS = 59
+    OVERLAY_FADES = 60
+    UNUSED_61 = 61
+    PHYSICS_LEVEL = 62  # used by orange_box_x360 for L4D2 maps
+    UNUSED_63 = 63
+
+
+LumpHeader = source.LumpHeader
+
+# a rough map of the relationships between lumps:
+
+#                     /-> SurfEdge -> Edge -> Vertex
+# Leaf -> Node -> Face -> Plane
+#                     \-> DisplacementInfo -> DisplacementVertex
+
+# ClipPortalVertices are AreaPortal geometry [citation neeeded]
+
+
+# classes for special lumps, in alphabetical order:
+# TODO: GameLump_SPRPv10
+
+
+# {"LUMP_NAME": {version: LumpClass}}
+BASIC_LUMP_CLASSES = sdk_2013.BASIC_LUMP_CLASSES.copy()
+
+LUMP_CLASSES = sdk_2013.LUMP_CLASSES.copy()
+
+SPECIAL_LUMP_CLASSES = sdk_2013.SPECIAL_LUMP_CLASSES.copy()
+
+GAME_LUMP_HEADER = sdk_2013.GAME_LUMP_HEADER
+
+# {"lump": {version: SpecialLumpClass}}
+GAME_LUMP_CLASSES = {"sprp": sdk_2013.GAME_LUMP_CLASSES["sprp"].copy()}
+GAME_LUMP_CLASSES["sprp"].pop(10)
+
+
+methods = [*sdk_2013.methods]